--- a/keywords_format.go
+++ b/keywords_format.go
@@ -1,7 +1,6 @@
 package jsonschema
 
 import (
-	// "encoding/json"
 	"fmt"
 	"net"
 	"net/mail"