As a continuation to PR #54, I attempted to see what would it take to implement a type system. It turned out not bad. The patch implements json type system, that maps any possible Go type to json type: `null`, `boolean`, `integer`, `number`, `string`, `object`, `array`. It allows validators not to care about specific Go type and their differences. For example, a validator can type assert `object` and read data without needing to differentiate between `map` and `struct`.

The most important part of the patch is a new file `types.go`. The rest is primarily patching existing validators. The new code needs testing, which I'll do if you consider merging.

I benchmarked the current master and this branch using benchmarks from PR #56. It has a significant performance degradation running `type` validator. I haven't checked exactly where most of the time is spent, but is worth to see if we can optimize it.
```
// master
BenchmarkAdditionalItems/sample_size_1-8         	 3000000	       411 ns/op	      52 B/op	       5 allocs/op
BenchmarkAdditionalItems/sample_size_10-8        	 1000000	      2261 ns/op	     232 B/op	      32 allocs/op
BenchmarkAdditionalItems/sample_size_100-8       	  100000	     21225 ns/op	    2248 B/op	     302 allocs/op
BenchmarkAdditionalItems/sample_size_1000-8      	   10000	    241823 ns/op	   27448 B/op	    3902 allocs/op
BenchmarkAdditionalProperties/sample_size_1-8    	 3000000	       571 ns/op	      54 B/op	       5 allocs/op
BenchmarkAdditionalProperties/sample_size_10-8   	  300000	      4278 ns/op	     256 B/op	      32 allocs/op
BenchmarkAdditionalProperties/sample_size_100-8  	   10000	    105712 ns/op	    2428 B/op	     302 allocs/op
BenchmarkAdditionalProperties/sample_size_1000-8 	     200	   8885443 ns/op	   26613 B/op	    3002 allocs/op
BenchmarkConst/sample_size_1-8                   	 1000000	      1141 ns/op	     720 B/op	      13 allocs/op
BenchmarkConst/sample_size_10-8                  	  200000	      6168 ns/op	    2134 B/op	      68 allocs/op
BenchmarkConst/sample_size_100-8                 	   30000	     58875 ns/op	   20591 B/op	     616 allocs/op
BenchmarkConst/sample_size_1000-8                	    2000	    643018 ns/op	  267220 B/op	    6037 allocs/op
BenchmarkContains/sample_size_1-8                	 3000000	       496 ns/op	     208 B/op	       3 allocs/op
BenchmarkContains/sample_size_10-8               	  200000	      9214 ns/op	    3170 B/op	      85 allocs/op
BenchmarkContains/sample_size_100-8              	   20000	     98430 ns/op	   33511 B/op	    1094 allocs/op
BenchmarkContains/sample_size_1000-8             	    2000	   1010793 ns/op	  336124 B/op	   10994 allocs/op
BenchmarkDependencies/sample_size_1-8            	 5000000	       404 ns/op	      38 B/op	       4 allocs/op
BenchmarkDependencies/sample_size_10-8           	 2000000	       602 ns/op	      38 B/op	       4 allocs/op
BenchmarkDependencies/sample_size_100-8          	 1000000	      2200 ns/op	      38 B/op	       4 allocs/op
BenchmarkDependencies/sample_size_1000-8         	   50000	     26662 ns/op	      38 B/op	       4 allocs/op
BenchmarkEnum/sample_size_1-8                    	 3000000	       415 ns/op	     176 B/op	       2 allocs/op
BenchmarkEnum/sample_size_10-8                   	  300000	      5131 ns/op	    1657 B/op	      42 allocs/op
BenchmarkEnum/sample_size_100-8                  	   30000	     46125 ns/op	   15306 B/op	     483 allocs/op
BenchmarkEnum/sample_size_1000-8                 	    3000	    455323 ns/op	  152206 B/op	    4983 allocs/op
BenchmarkMaximum/sample_size_1-8                 	30000000	        46.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_10-8                	30000000	        46.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_100-8               	30000000	        46.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_1000-8              	30000000	        46.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_1-8                 	30000000	        46.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_10-8                	30000000	        46.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_100-8               	30000000	        47.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_1000-8              	30000000	        46.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_1-8        	30000000	        46.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_10-8       	30000000	        46.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_100-8      	30000000	        45.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_1000-8     	30000000	        46.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_1-8        	30000000	        45.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_10-8       	30000000	        46.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_100-8      	30000000	        44.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_1000-8     	30000000	        46.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_1-8                	30000000	        46.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_10-8               	30000000	        47.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_100-8              	30000000	        46.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_1000-8             	30000000	        46.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_1-8                	30000000	        45.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_10-8               	30000000	        46.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_100-8              	30000000	        45.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_1000-8             	30000000	        45.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_1-8               	30000000	        47.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_10-8              	30000000	        52.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_100-8             	20000000	       115 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_1000-8            	 2000000	       614 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_1-8               	30000000	        48.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_10-8              	30000000	        52.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_100-8             	20000000	       117 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_1000-8            	 2000000	       608 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_1-8           	30000000	        45.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_10-8          	30000000	        46.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_100-8         	30000000	        45.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_1000-8        	30000000	        45.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_1-8           	30000000	        45.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_10-8          	30000000	        45.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_100-8         	30000000	        44.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_1000-8        	30000000	        45.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_1-8              	30000000	        44.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_10-8             	30000000	        46.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_100-8            	30000000	        46.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_1000-8           	30000000	        46.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkPattern/sample_size_1-8                 	 5000000	       245 ns/op	     152 B/op	       2 allocs/op
BenchmarkPattern/sample_size_10-8                	 3000000	       433 ns/op	     160 B/op	       2 allocs/op
BenchmarkPattern/sample_size_100-8               	 1000000	      2184 ns/op	     256 B/op	       2 allocs/op
BenchmarkPattern/sample_size_1000-8              	  100000	     19968 ns/op	    1168 B/op	       2 allocs/op
BenchmarkType/sample_size_1-8                    	 1000000	      2233 ns/op	     150 B/op	      19 allocs/op
BenchmarkType/sample_size_10-8                   	  100000	     23057 ns/op	    1360 B/op	     181 allocs/op
BenchmarkType/sample_size_100-8                  	    5000	    261737 ns/op	   14396 B/op	    1801 allocs/op
BenchmarkType/sample_size_1000-8                 	     500	   2859282 ns/op	  159515 B/op	   18001 allocs/op


// json type system
BenchmarkAdditionalItems/sample_size_1-8         	 3000000	       400 ns/op	      52 B/op	       5 allocs/op
BenchmarkAdditionalItems/sample_size_10-8        	 1000000	      2413 ns/op	     336 B/op	      41 allocs/op
BenchmarkAdditionalItems/sample_size_100-8       	  100000	     22550 ns/op	    3216 B/op	     401 allocs/op
BenchmarkAdditionalItems/sample_size_1000-8      	    5000	    259886 ns/op	   39216 B/op	    4901 allocs/op
BenchmarkAdditionalProperties/sample_size_1-8    	  300000	      4163 ns/op	    4102 B/op	      16 allocs/op
BenchmarkAdditionalProperties/sample_size_10-8   	  100000	     14283 ns/op	    5024 B/op	      88 allocs/op
BenchmarkAdditionalProperties/sample_size_100-8  	   10000	    157243 ns/op	   14403 B/op	     808 allocs/op
BenchmarkAdditionalProperties/sample_size_1000-8 	     100	  10425760 ns/op	  110580 B/op	    8008 allocs/op
BenchmarkConst/sample_size_1-8                   	  500000	      2403 ns/op	    1520 B/op	      28 allocs/op
BenchmarkConst/sample_size_10-8                  	  100000	     12383 ns/op	    4766 B/op	     131 allocs/op
BenchmarkConst/sample_size_100-8                 	   10000	    123338 ns/op	   45159 B/op	    1137 allocs/op
BenchmarkConst/sample_size_1000-8                	    1000	   1390451 ns/op	  580925 B/op	   11082 allocs/op
BenchmarkContains/sample_size_1-8                	 1000000	      1009 ns/op	     392 B/op	       6 allocs/op
BenchmarkContains/sample_size_10-8               	  100000	     14293 ns/op	    5291 B/op	     150 allocs/op
BenchmarkContains/sample_size_100-8              	   10000	    155046 ns/op	   55881 B/op	    1879 allocs/op
BenchmarkContains/sample_size_1000-8             	    1000	   1635063 ns/op	  568191 B/op	   18980 allocs/op
BenchmarkDependencies/sample_size_1-8            	 3000000	       532 ns/op	      86 B/op	       7 allocs/op
BenchmarkDependencies/sample_size_10-8           	 1000000	      1645 ns/op	     374 B/op	      25 allocs/op
BenchmarkDependencies/sample_size_100-8          	  100000	     13244 ns/op	    3254 B/op	     205 allocs/op
BenchmarkDependencies/sample_size_1000-8         	   10000	    119900 ns/op	   32054 B/op	    2005 allocs/op
BenchmarkEnum/sample_size_1-8                    	 2000000	       933 ns/op	     360 B/op	       5 allocs/op
BenchmarkEnum/sample_size_10-8                   	  200000	      7833 ns/op	    2898 B/op	      77 allocs/op
BenchmarkEnum/sample_size_100-8                  	   20000	     69402 ns/op	   26644 B/op	     839 allocs/op
BenchmarkEnum/sample_size_1000-8                 	    2000	    681553 ns/op	  264422 B/op	    8489 allocs/op
BenchmarkMaximum/sample_size_1-8                 	30000000	        40.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_10-8                	30000000	        41.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_100-8               	30000000	        40.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaximum/sample_size_1000-8              	30000000	        41.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_1-8                 	30000000	        41.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_10-8                	30000000	        40.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_100-8               	30000000	        41.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinimum/sample_size_1000-8              	30000000	        41.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_1-8        	30000000	        40.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_10-8       	30000000	        40.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_100-8      	30000000	        41.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMaximum/sample_size_1000-8     	30000000	        41.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_1-8        	30000000	        40.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_10-8       	30000000	        40.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_100-8      	30000000	        41.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkExclusiveMinimum/sample_size_1000-8     	30000000	        41.3 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_1-8                	30000000	        40.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_10-8               	30000000	        40.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_100-8              	30000000	        39.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxItems/sample_size_1000-8             	30000000	        40.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_1-8                	30000000	        41.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_10-8               	30000000	        40.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_100-8              	30000000	        40.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinItems/sample_size_1000-8             	30000000	        39.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_1-8               	30000000	        43.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_10-8              	30000000	        47.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_100-8             	20000000	       106 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxLength/sample_size_1000-8            	 3000000	       565 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_1-8               	30000000	        42.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_10-8              	30000000	        47.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_100-8             	20000000	       104 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinLength/sample_size_1000-8            	 3000000	       580 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_1-8           	30000000	        49.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_10-8          	30000000	        48.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_100-8         	30000000	        52.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkMaxProperties/sample_size_1000-8        	20000000	        51.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_1-8           	30000000	        49.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_10-8          	30000000	        49.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_100-8         	30000000	        48.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkMinProperties/sample_size_1000-8        	30000000	        51.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_1-8              	30000000	        43.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_10-8             	30000000	        42.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_100-8            	30000000	        42.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkMultipleOf/sample_size_1000-8           	30000000	        42.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkPattern/sample_size_1-8                 	10000000	       229 ns/op	     152 B/op	       2 allocs/op
BenchmarkPattern/sample_size_10-8                	 3000000	       390 ns/op	     160 B/op	       2 allocs/op
BenchmarkPattern/sample_size_100-8               	 1000000	      1840 ns/op	     256 B/op	       2 allocs/op
BenchmarkPattern/sample_size_1000-8              	  100000	     19207 ns/op	    1168 B/op	       2 allocs/op
BenchmarkType/sample_size_1-8                    	  200000	      6459 ns/op	    2460 B/op	      40 allocs/op
BenchmarkType/sample_size_10-8                   	   30000	     43821 ns/op	    6612 B/op	     364 allocs/op
BenchmarkType/sample_size_100-8                  	    3000	    408302 ns/op	   48407 B/op	    3604 allocs/op
BenchmarkType/sample_size_1000-8                 	     300	   4176650 ns/op	  489146 B/op	   36004 allocs/op
```